requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
aiohttp==3.9.1
Pillow==10.1.0
requests==2.31.0
cryptography==41.0.7
Flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==3.0.1
yt-dlp==2023.12.30
opuslib==3.0.1
ghapi==1.0.4
wikipedia==1.4.0
wolframalpha==5.0.0
psutil==5.9.6
psutil-windows==5.9.6
wavelink==3.0.0
